chore: Add AB flag storage and enable testWorkerLogic flag
