+++ src/components/Synthetics/PositionList/PositionList.tsx
@@ -1,6 +1,7 @@
 import { Trans, t } from "@lingui/macro";
 import { memo, useCallback, useEffect, useRef, useState } from "react";
 
+import { getIsFlagEnabled } from "config/ab";
 import { useIsPositionsLoading, usePositionsInfoData } from "context/SyntheticsStateContext/hooks/globalsHooks";
 import { usePositionEditorPositionState } from "context/SyntheticsStateContext/hooks/positionEditorHooks";
 import { selectAccount, selectChainId } from "context/SyntheticsStateContext/selectors/globalSelectors";
@@ -14,8 +15,8 @@ import { getByKey } from "lib/objects";
 import PositionShare from "components/Exchange/PositionShare";
 import { OrderEditorContainer } from "components/OrderEditorContainer/OrderEditorContainer";
 import { PositionItem } from "components/Synthetics/PositionItem/PositionItem";
-import { useMetrics } from "context/MetricsContext/MetricsContext";
 import { DATA_LOAD_TIMEOUT_FOR_METRICS } from "config/ui";
+import { useMetrics } from "context/MetricsContext/MetricsContext";
 import { useLatest } from "react-use";
 
 type Props = {
@@ -47,6 +48,7 @@ export function PositionList(p: Props) {
 
   const metricsRef = useLatest(metrics);
   const metricsTimeout = useRef<NodeJS.Timeout>();
+  const requestIdRef = useRef<string>();
 
   useEffect(() => {
     if (metricsTimeout.current) {
@@ -55,9 +57,14 @@ export function PositionList(p: Props) {
 
     metricsRef.current.startTimer("positionsList");
 
+    requestIdRef.current = Date.now().toString() + "_" + Math.trunc(Math.random() * 1000000).toString();
     metricsRef.current.sendMetric({
       event: "positionsListLoad.started",
       isError: false,
+      data: {
+        testWorkersLogic: getIsFlagEnabled("testWorkerLogic"),
+        requestId: requestIdRef.current,
+      },
     });
 
     metricsTimeout.current = setTimeout(() => {
@@ -66,6 +73,10 @@ export function PositionList(p: Props) {
         message: "Positions list was not loaded",
         isError: true,
         time: metricsRef.current.getTime("positionsList"),
+        data: {
+          testWorkersLogic: getIsFlagEnabled("testWorkerLogic"),
+          requestId: requestIdRef.current!,
+        },
       });
     }, DATA_LOAD_TIMEOUT_FOR_METRICS);
   }, [metricsRef]);
@@ -78,6 +89,10 @@ export function PositionList(p: Props) {
         event: "positionsListLoad.success",
         isError: false,
         time: metrics.getTime("positionsList"),
+        data: {
+          testWorkersLogic: getIsFlagEnabled("testWorkerLogic"),
+          requestId: requestIdRef.current!,
+        },
       });
       setIsLoaded(true);
     }
